reorganise DID methods, small tweaks to lpdid
